# Conflicts:
#	climaf/cache.py
#	doc/news.rst
#	doc/std_operators.rst

+++ climaf/driver.py
@@ -82,12 +82,13 @@ def capply_script (script_name, *operands, **parameters):
     if (isinstance(operands[0],classes.cens) and script.flags.commuteWithEnsemble) :
         # Must iterate on members
         reps=[]
-        for member,label in zip(operands[0].members,operands[0].labels) :
+        for label in operands[0].order:
+            member=operands[0][label]
             clogger.debug("processing member "+`member`)
             params=parameters.copy()
             params["member_label"]=label
             reps.append(maketree(script_name, script, member, *opscopy, **params))
-        return(classes.cens(operands[0].labels,*reps))
+        return(classes.cens(dict(zip(operands[0].order,reps))))
     else: 
         return(maketree(script_name, script, *operands, **parameters))
             
@@ -103,12 +104,16 @@ def maketree(script_name, script, *operands, **parameters):
                 rep.variable=script.outputs['']%defaultVariable
             else:
                 rep.variable=script.outputs['']
+            template=Template(rep.variable)
+            rep.variable=template.substitute(parameters)
         else :
             son=classes.scriptChild(rep,outname)
             if "%s" in script.outputs[outname] :
                 son.variable=script.outputs[outname]%defaultVariable
             else:
                 son.variable=script.outputs[outname]
+            template=Template(son.variable)
+            son.variable=template.substitute(parameters)
             rep.outputs[outname]=son
             setattr(rep,outname,son)
     return rep
@@ -181,6 +186,7 @@ def ceval(cobject, userflags=None, format="MaskedArray",
                 (userflags.canAggregateTime or ds.periodHasOneFile()) and \
                 (userflags.canAlias         or ds.hasExactVariable()) and \
                 (userflags.canMissing       or ds.missingIsOK())      and \
+                (ds.hasOneMember()) and \
                 #(userflags.doSqueezeMembers or ds.hasOneMember()) and 
                 (format == 'file')) :
                 clogger.debug("Delivering file set or sets is OK for the target use")
@@ -196,8 +202,12 @@ def ceval(cobject, userflags=None, format="MaskedArray",
                 ## clogger.debug(" Done with subsetting and caching data files")
                 ## cstore(extract) # extract should include the dataset def
                 ## return ceval(extract,userflags,format)
-                clogger.debug("Fetching/selection/aggregation is done using an external script for now - TBD")
-                extract=capply('select',ds)
+                if ds.hasOneMember() :
+                    clogger.debug("Fetching/selection/aggregation is done using an external script for now - TBD")
+                    extract=capply('select',ds)
+                else :
+                    clogger.debug("On multi-member datafiles , fetching/selection/aggregation is done using select_member - TBD")
+                    extract=capply('select_member',ds)
                 if extract is None : raise Climaf_Driver_Error("Cannot access dataset" + `ds`)
                 rep=ceval(extract,userflags=userflags,format=format)
                 userflags.unset_selectors()
@@ -209,12 +219,13 @@ def ceval(cobject, userflags=None, format="MaskedArray",
             #   then assume it can also select on time and provide it with the address
             #   else : fetch the relevant selection of the data, and store it in cache
             clogger.debug("Dataset is remote " )
-            if (userflags.canOpenDap and format == 'file' ) :
+            if (userflags.canOpendap and format == 'file' ) :
                 clogger.debug("But user can OpenDAP " )
                 cdedent()
                 return(ds.adressOf())
             else :
                 clogger.debug("Must remote read and cache " )
+                raise Climaf_Driver_Error("Cannot acces dataset %s"%`cobject`)
                 rep=ceval(capply('remote_select',ds),userflags=userflags,format=format)
                 userflags.unset_selectors()
                 cdedent()
@@ -312,11 +323,11 @@ def ceval(cobject, userflags=None, format="MaskedArray",
             if ( format == 'file' ) : return (file)
             else : return cread(file)
         elif isinstance(cobject,classes.cens) :
-            rep=[]
-            for member in cobject.members :
-                rep.append(ceval(member,copy.copy(userflags),format,deep,recurse_list=recurse_list))
-            if (format=="file") : return(reduce(lambda x,y : x+" "+y, rep))
-            else : return rep
+            d=dict()
+            for member in cobject.order :
+                d[member]=ceval(cobject[member],copy.copy(userflags),format,deep,recurse_list=recurse_list)
+            if (format=="file") : return(reduce(lambda x,y : x+" "+y, [ d[m] for m in cobject.order ]))
+            else : return d
         else :
             raise Climaf_Driver_Error("Internal logic error")
     elif isinstance(cobject,str) :
@@ -375,8 +386,8 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
             filevar,scale,offset,units,filenameVar,missing=op.alias
             if scriptCall.flags.canAlias and "," not in varOf(op) :
                 #if script=="select" and ((varOf(op) != filevar) or scale != 1.0 or offset != 0.) :
-                subdict["alias"]="%s,%s,%.4g,%.4g"%(varOf(op),filevar,scale,offset)
-                subdict["var"]=filevar
+                subdict["var"]=Template(filevar).safe_substitute(op.kvp)
+                subdict["alias"]="%s,%s,%.4g,%.4g"%(varOf(op),subdict["var"],scale,offset)
             if units : subdict["units"]=units 
             if scriptCall.flags.canMissing and missing :
                 subdict["missing"]=missing
@@ -386,7 +397,7 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
                     "Script %s 's input #%s cannot accept ensemble %s"\
                         %(scriptCall.script,0,`op`))
             #subdict["labels"]=r'"'+reduce(lambda x,y : "'"+x+"' '"+y+"'", op.labels)+r'"'
-            subdict["labels"]=reduce(lambda x,y : x+"$"+y, op.labels)
+            subdict["labels"]=reduce(lambda x,y : x+"$"+y, op.order)
         if op : 
             per=timePeriod(op)
             if per and not per.fx and str(per) != "" and scriptCall.flags.canSelectTime:
@@ -413,8 +424,8 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
                 filevar,scale,offset,units,filenameVar,missing =op.alias
                 if ((varOf(op) != filevar) or (scale != 1.0) or (offset != 0.)) and \
                         "," not in varOf(op):
-                    subdict["alias_%d"%i]="%s %s %f %f"%(varOf(op),filevar,scale,offset)
-                    subdict["var_%d"%i]=filevar
+                    subdict["var_%d"%i]=Template(filevar).safe_substitute(op.kvp)
+                    subdict["alias_%d"%i]="%s %s %f %f"%(varOf(op),subdict["var_%d"%i],scale,offset)
                 if units : subdict["units_%d"%i]=units 
                 if missing : subdict["missing_%d"%i]=missing
                 # Provide period selection if script allows for
@@ -476,18 +487,17 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
     # Allowing for some formal parameters to be missing in the actual call:
     #
     # Discard remaining substrings looking like :
-    #  some_word='"${some_keyword}"'     , or:
-    #  '"${some_keyword}"'
+    #  some_word='"${some_keyword}"' , or simply : '"${some_keyword}"'
     template=re.sub(r'(\w*=)?(\'\")?\$\{\w*\}(\"\')?',r"",template)
     #
     # Discard remaining substrings looking like :
-    #  some_word=${some_keyword}          , or
-    #  ${some_keyword}
+    #  some_word=${some_keyword}  or  simply : ${some_keyword}
     template=re.sub(r"(\w*=)?\$\{\w*\}",r"",template)
     #
     # Link the fixed fields needed by the script/operator
     if script.fixedfields is not None :
-        subdict_ff=dict()
+        #subdict_ff=dict()
+        subdict_ff=scriptCall.parameters.copy()
         subdict_ff["model"]=modelOf(scriptCall.operands[0])
         subdict_ff["simulation"]=simulationOf(scriptCall.operands[0])
         subdict_ff["project"]=projectOf(scriptCall.operands[0])
@@ -497,24 +507,22 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
         for ll,lt in l:
             #Replace input data placeholders with filenames for fixed fields
             template_ff_target=Template(lt).substitute(subdict_ff)
-            # symlink
+            # symlink if needed
             files_exist[ll]=False
-            if os.path.isfile(ll):      
-                files_exist[ll]=True
-            else:
-                os.system("ln -s "+template_ff_target+" "+ll)   
-            
-    # Launch script using command, and check termination 
-    #command="PATH=$PATH:"+operators.scriptsPath+template+fileVariables
-    #command="echo '\n\nstdout and stderr of script call :\n\t "+template+\
-    #         "\n\n'> scripts.out  ; "+ template+ " >> scripts.out 2>&1"
-
+            if os.path.isfile(ll): files_exist[ll]=True
+            else: os.system("ln -s "+template_ff_target+" "+ll)   
+    #
     tim1=time.time()
     clogger.info("Launching command:"+template)
     #
-    command=subprocess.Popen(template, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, shell=True)
-    repcom=command.wait()
-    #
+    try :
+        command=subprocess.Popen(template, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, shell=True)
+        repcom=command.wait()
+        #
+    except:
+        clogger.error("Subprocess was interrupted")
+        repcom=1
+    
     logfile=open('last.out', 'w')
     logfile.write("\n\nstdout and stderr of script call :\n\t "+template+"\n\n")
     command_std=""
@@ -524,13 +532,12 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
         if script.outputFormat=="txt" : 
             sys.stdout.write(line)
     logfile.close()
-    
+
     # Clean fixed fields symbolic links
-    if script.fixedfields is not None :
-        l=script.fixedfields  #return paths: (linkname, targetname)
-        for ll,lt in l:
-            if files_exist[ll] == False:
-                os.system("rm -f "+ll)  
+    # (linkname, targetname)
+    if script.fixedfields :
+        for ll,lt in script.fixedfields:
+            if files_exist[ll] == False: os.system("rm -f "+ll)  
 
     if ( repcom == 0 ):
         if script.outputFormat not in operators.none_formats :
@@ -546,9 +553,9 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
                 #    set_variable(subdict["out_"+output], output, 'file')
             if ok : 
                 duration=time.time() - tim1
-                print("Done in %.1f s with script computation for %s "%\
-                          (duration,`scriptCall`),file=sys.stderr)
-                clogger.debug("Done in %.1f s with script computation for "
+                #print("Done in %.1f s with script computation for %s "%\
+                #          (duration,`scriptCall`),file=sys.stderr)
+                clogger.info("Done in %.1f s with script computation for "
                               "%s (command was :%s )"%\
                                   (duration,`scriptCall`,template))
                 return main_output_filename
@@ -564,10 +571,6 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
         clogger.error("Last lines of script output:\n"+comm2.stdout.read())
         raise Climaf_Driver_Error("Script failure for : %s. More details either in file "
                                   "./last.out or by re-runing with clog(\"debug\")" %template)
-        #raise Climaf_Driver_Error("See above (or scripts.out <=> clog('debug')) for analyzing "
-        #                          "script call failure for : %s "%template)
-
-
 
 def timePeriod(cobject) :
     """ Returns a time period for a CliMAF object : if object is a dataset, returns
@@ -582,7 +585,7 @@ def timePeriod(cobject) :
         return timePeriod(cobject.father)
     elif isinstance(cobject,classes.cens) :
         clogger.debug("for now, timePeriod logic for 'cens' objet is basic (1st member)- TBD")
-        return timePeriod(cobject.members[0])
+        return timePeriod(cobject.values()[0])
     else : return None #clogger.error("unkown class for argument "+`cobject`)
                   
 def domainOf(cobject) :
@@ -606,7 +609,7 @@ def domainOf(cobject) :
         return domainOf(cobject.father)
     elif isinstance(cobject,classes.cens) :
         clogger.debug("for now, domainOf logic for 'cens' objet is basic (1st member)- TBD")
-        return domainOf(cobject.members[0])
+        return domainOf(cobject.values()[0])
     elif cobject is None : return "none"
     else : clogger.error("Unkown class for argument "+`cobject`)
                   
@@ -625,7 +628,7 @@ def attributeOf(cobject,attrib) :
         val=getattr(cobject,attrib,None) 
         if val is not None : return val
         else : return(cobject.kvp.get(attrib))
-    elif isinstance(cobject,classes.cens) : return attributeOf(cobject.members[0],attrib)
+    elif isinstance(cobject,classes.cens) : return attributeOf(cobject.values()[0],attrib)
     elif getattr(cobject,attrib,None) : return getattr(cobject,attrib) 
     elif isinstance(cobject,classes.ctree) :
         clogger.debug("for now, varOf logic is basic (1st operand) - TBD")
@@ -655,6 +658,7 @@ def ceval_select(includer,included,userflags,format,deep,derived_list,recurse_li
             return(cache.rename(objfile,crs))
         else :
             clogger.critical("Cannot evaluate "+`extract`)
+            exit()
     else :
         clogger.error("Can yet process only files - TBD")
 
@@ -1114,7 +1118,7 @@ def efile(obj, filename, forced=False) :
         filename (str) : output filename. It will include a field for each 
          ensemble's member, with a variable name suffixed by the member
          label (e.g. : tas_CNRM-CM, tas_IPSL-CM... ) (more formally : 
-         'var(obj.members[n])'_'obj.labels[n]')
+         'var(obj.order[n])'_'obj.ens[order[n]]')
 
         forced (logical, optional) : if True, CliMAF will override the file
          'filename' if it already exists 
@@ -1129,7 +1133,8 @@ def efile(obj, filename, forced=False) :
             else:
                 raise Climaf_Driver_Error("File '%s' already exists: use 'forced=True' to override it" %filename)
                 
-        for memb,lab in zip(obj.members,obj.labels):
+        for lab in obj.order:
+            memb=obj[lab]
             ffile=cfile(memb)
             
             f = tempfile.NamedTemporaryFile(suffix=".nc")
